- use subcommand folders - simplify and combine true/false, public/private commands to parent options
